fix: improved scaling
Resize the canvas image depending on scale factor instead of transform: scale() the UI. This allows buffs to be modified when scaled and generally cleans up scaling nicely.

It can still get cut off on the right side - if that happens make the Alt1 app window wider before minimizing. If you go to wide for some reason the canvas... shrinks? Not sure why. But this is better than things were at least.